+++ src/lib/utils/options/tsdoc-defaults.ts
@@ -55,4 +55,5 @@ export const modifierTags = [
     "@event",
     "@overload",
     "@namespace",
+    "@interface",
 ] as const;